Update TwoCityScheduling.py
There are 2N people a company is planning to interview. The cost of flying the i-th person to city A is costs[i][0], and the cost of flying the i-th person to city B is costs[i][1]. A function to return the minimum cost to fly every person to a city such that exactly N people arrive in each city.